Rename compact() to formatCompactedComparison()
The name of the function is strange. Although it does compact the strings, it
actually might not compact the strings if canBeCompacted returns false. So naming this
function compact hides  the  side  effect  of  the  error  check.  Notice  also  that  the  function
returns a formatted message, not just the compacted strings. So the name of the function
should really be formatCompactedComparison.